SyntacticSuite: split by whether quasi or dotty
